# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server API e Websocket per il progetto HermesChat. Il server è stato sviluppato in Go con il framework Gin e utilizza MongoDB come database.",
        "title": "HermesChat API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.hermeschat.it:8090",
    "basePath": "/",
    "paths": {
...
        "/getFiles": {
            "post": {
                "description": "Scarica i file dagli url specificati e li invia al client",
                "produces": [
                    "application/json"
                ],
                "summary": "Scarica i file",
                "parameters": [
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "Url dei file da scaricare",
                        "name": "urls",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "description": "ChatId delle chat in cui si trovano i file",
                        "name": "chatId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/login": {
            "post": {
                "description": "Effettua il login di un utente e crea un token JWT salvato in un cookie http-only",
                "produces": [
                    "application/json"
                ],
                "summary": "Login utente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username dell'utente",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password dell'utente",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/sendFile": {
            "post": {
                "description": "Salva il file nella cartella del server dell'utente loggato e invia il file all'utente specificato",
                "produces": [
                    "application/json"
                ],
                "summary": "Invia un file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indice dell'utente loggato",
                        "name": "index",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username dell'utente a cui inviare il file",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "File da inviare",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/sendFriendRequest": {
            "post": {
                "description": "Invia una richiesta di amicizia all'utente specificato dall'utente loggato",
                "produces": [
                    "application/json"
                ],
                "summary": "Invia richiesta di amicizia",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Indice dell'utente loggato",
                        "name": "index",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Username dell'utente a cui inviare la richiesta di amicizia",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "Effettua la registrazione di un utente e crea un codice OTP che verrà inviato via email",
                "produces": [
                    "application/json"
                ],
                "summary": "Registrazione di un utente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username dell'utente",
                        "name": "username",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password dell'utente",
                        "name": "password",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Email dell'utente",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}